Moved out definition of tables related to Comments in their own models file
