+++ source/app/datamgmt/case/case_assets_db.py
@@ -29,12 +29,11 @@ from app.datamgmt.filtering import get_filtered_data
 from app.datamgmt.states import update_assets_state
 from app.models.models import AnalysisStatus
 from app.models.models import CaseStatus
-from app.models.models import AssetComments
 from app.models.models import AssetsType
 from app.models.models import CaseAssets
 from app.models.models import CaseEventsAssets
 from app.models.cases import Cases
-from app.models.models import Comments
+from app.models.comments import Comments, AssetComments
 from app.models.models import CompromiseStatus
 from app.models.iocs import Ioc
 from app.models.models import IocAssetLink
