+++ source/app/datamgmt/case/case_iocs_db.py
@@ -29,9 +29,8 @@ from app.models.cases import Cases
 from app.models.cases import CasesEvent
 from app.models.models import Client
 from app.models.models import CaseAssets
-from app.models.models import Comments
+from app.models.comments import Comments, IocComments
 from app.models.iocs import Ioc
-from app.models.models import IocComments
 from app.models.models import IocType
 from app.models.iocs import Tlp
 from app.models.authorization import User
