+++ source/app/datamgmt/case/case_notes_db.py
@@ -21,11 +21,10 @@ from app import db
 from app.iris_engine.access_control.iris_user import iris_current_user
 from app.datamgmt.manage.manage_attribute_db import get_default_custom_attributes
 from app.datamgmt.states import update_notes_state
-from app.models.models import Comments
+from app.models.comments import Comments, NotesComments
 from app.models.models import NoteDirectory
 from app.models.models import NoteRevisions
 from app.models.models import Notes
-from app.models.models import NotesComments
 from app.models.models import NotesGroup
 from app.models.models import NotesGroupLink
 from app.models.authorization import User
