+++ source/app/datamgmt/case/case_tasks_db.py
@@ -29,8 +29,7 @@ from app.datamgmt.states import update_tasks_state
 from app.models.models import CaseTasks
 from app.models.models import TaskAssignee
 from app.models.cases import Cases
-from app.models.models import Comments
-from app.models.models import TaskComments
+from app.models.comments import Comments, TaskComments
 from app.models.models import TaskStatus
 from app.models.authorization import User
 from app.models.pagination_parameters import PaginationParameters
