+++ source/app/schema/marshables.py
@@ -69,7 +69,7 @@ from app.models.models import CaseTasks
 from app.models.cases import Cases
 from app.models.cases import CasesEvent
 from app.models.models import Client
-from app.models.models import Comments
+from app.models.comments import Comments
 from app.models.models import Contact
 from app.models.models import DataStoreFile
 from app.models.models import EventCategory
